requirements.txt
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
transformers
tokenizers
sentencepiece
shortuuid
accelerate
peft
bitsandbytes
pydantic 
markdown2[all]
numpy 
scikit-learn
gradio
gradio_client
requests
httpx
uvicorn
protobuf
fastapi
einops
einops-exts
timm
opencv-python
fvcore
# these lib are required for the lmms-eval
black
datasets
evaluate>=0.4.0
jsonlines
numexpr
pybind11>=2.6.2
pytablewriter
rouge-score>=0.0.4
sacrebleu>=1.5.0
scikit-learn>=0.24.1
sqlitedict
openai>=1.0.0
pycocoevalcap
tqdm-multiprocess
zstandard
pyyaml
sympy
mpmath
Jinja2
openpyxl
Levenshtein
hf_transfer
tenacity
transformers-stream-generator
tiktoken
pre-commit
ninja